node v4.1.2 works! also Rob fixes on cert chain
Robby checked that a certificate in the chain
was missing, and browsers are not strict about it, while camo is.

tested with older node packets, but it's not working: they're still
failing with error [1]. 4.1.2 fixed it and we just have to
wait until the new build gets into the repositories.

[1]: https://gist.github.com/alfateam123/d4b9d6556d5d0331f92a